Clean stats file helper
