+++ app/services/stats/needs/done_not_reachable.rb
@@ -27,7 +27,7 @@ module Stats::Needs
     end
 
     def secondary_count
-      filtered_main_query.status_done_not_reachable.size
+      @secondary_count ||= filtered_main_query.status_done_not_reachable.size
     end
 
     private
