+++ app/services/stats/needs/taking_care.rb
@@ -26,7 +26,7 @@ module Stats::Needs
     end
 
     def secondary_count
-      filtered_main_query.status_taking_care.size
+      @secondary_count ||= filtered_main_query.status_taking_care.size
     end
 
     private
